+++ Runtime/SoundEvent.cs
@@ -22,18 +22,18 @@ namespace PixelDust.Audiophile
             SoundManager.StopSound(Data.SoundId);
         }
 
-        public AudiophilePlayResult PlayAt(Vector3 position, ulong delay = 0, string overrideId = null)
+        public AudiophilePlayResult PlayAt(Vector3 position, float delay = 0, string overrideId = null)
         {
             return SoundManager.ProcessSound(this, position, delay, overrideId);
         }
 
-        public AudiophilePlayResult PlayAt(Transform transform, ulong delay = 0, string overrideId = null)
+        public AudiophilePlayResult PlayAt(Transform transform, float delay = 0, string overrideId = null)
         {
             Vector3 position = transform != null ? transform.position : Vector3.zero;
             return PlayAt(position, delay, overrideId);
         }
 
-        public AudiophilePlayResult Play(ulong delay = 0, string overrideId = null)
+        public AudiophilePlayResult Play(float delay = 0, string overrideId = null)
         {
             return PlayAt(Vector3.zero, delay, overrideId);
         }
